@@ -1,10 +1,5 @@
 import { Injectable } from "@angular/core";
-import {
-	HttpRequest,
-	HttpHandler,
-	HttpEvent,
-	HttpInterceptor,
-} from "@angular/common/http";
+import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from "@angular/common/http";
 import { Observable } from "rxjs";
 import { UserService } from "src/app/shared/services/user.service";
 
