@@ -6,7 +6,7 @@ import { DataRowTransformationError } from "src/app/shared/model/data-row-transf
 import { catchError, map, Observable, of, startWith, switchMap } from "rxjs";
 import {DataConfigurationService} from "../../services/data-configuration.service";
 import {DataConfiguration} from "../../model/data-configuration";
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {environments} from "../../../../environments/environment";
 import { DataSetInfo } from "../../model/data-set-info";
 import { DataSetInfoService } from "../../../features/data-upload/services/data-set-info.service";
